requirements.txt
pyttsx3>=2.90
SpeechRecognition>=3.10.0
opencv-python>=4.8.0
numpy>=1.24.0
openai>=1.0.0
pywhatkit>=5.4
schedule>=1.2.0
pyautogui>=0.9.54
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
